Tests

+++ RedPitaya/Examples/C/regenerate.c
@@ -15,7 +15,7 @@ int main(int argc, char **argv){
 	int buff_size = 16384;
 	uint32_t posnow = 0;
 	uint32_t posold =0;
-	int zeros[16384];
+	float zeros[16384];
 	/* Print error, if rp_Init() function failed */
 	if(rp_Init() != RP_OK){
 		fprintf(stderr, "Rp api init failed!\n");
